+++ flutter_inappwebview_forge_android/android/src/main/kotlin/com/emirkanacar/flutter_inappwebview_forge_android/webview/in_app_webview/InAppWebView.kt
@@ -2506,9 +2506,17 @@ class InAppWebView : InputAwareWebView, InAppWebViewInterface {
       if (visibility != View.GONE) {
         super.onWindowVisibilityChanged(View.VISIBLE)
       }
+      if (visibility != View.GONE) {
+        postInvalidateOnAnimation()
+        requestLayout()
+      }
       return
     }
     super.onWindowVisibilityChanged(visibility)
+    if (visibility != View.GONE) {
+      postInvalidateOnAnimation()
+      requestLayout()
+    }
   }
 
   override fun getZoomScale(): Float = zoomScale